Выполнен перевод всех цен с int на BigDecimal.

+++ core-service/src/main/resources/db/migration/V2__orders.sql
@@ -1,7 +1,7 @@
 create table orders(
     id                  bigserial       primary key,
     username            varchar(255)    not null,
-    total_price         int             not null,
+    total_price         decimal(15, 2)  not null,
     address             varchar(255),
     phone               varchar(255),
     created_at          timestamp       default current_timestamp,
@@ -13,8 +13,8 @@ create table order_items(
     product_id          bigint          not null references products(id),
     order_id            bigint          not null references orders(id),
     quantity            int             not null,
-    price_per_product   int             not null,
-    price               int             not null,
+    price_per_product   decimal(15, 2)  not null,
+    price               decimal(15, 2)  not null,
     created_at          timestamp       default current_timestamp,
     updated_at          timestamp       default current_timestamp
 );
